refactor(backup): decompose the 1084-line restore loop into per-table resolvers behind a dispatch table, splitting the if/elif ladders into named identity resolvers (reconciling an incoming row against what the target already holds) and conflict resolvers (merge semantics applied just before insert), each returning an explicit INSERT or HANDLED outcome instead of a bare continue, move the restore into backend/core/backup/restore as state, stages, context, identity, conflicts and runner, relocate the BackupManager facade into the package and delete backend/core/backup_manager.py so its grandfathered file-size baseline can be removed from check_file_size.py outright, and split the remaining oversized functions so no file exceeds 500 lines and nothing inherits the old module's complexity exemption
