+++ backend/core/backup/__init__.py
@@ -1,1 +1,30 @@
-"""Backup and restore for a whole Nojoin installation."""
+"""Backup and restore for a whole Nojoin installation.
+
+The package is split by phase. ``format`` holds the archive contract, ``runtime`` the
+bindings the test harness substitutes, ``export`` builds an archive, and ``restore``
+takes one apart again. ``BackupManager`` is the facade the API and worker tasks use.
+"""
+
+from backend.core.backup.format import (
+    ARCHIVE_QUALITIES,
+    ARCHIVE_QUALITY_COMPRESSED,
+    ARCHIVE_QUALITY_ORIGINAL,
+    BACKUP_EXPORT_DIR,
+    BACKUP_FORMAT_VERSION,
+    RESTORE_LOCK_KEY,
+    RESTORE_LOCK_TTL_SECONDS,
+    RESTORE_STAGING_DIRNAME,
+)
+from backend.core.backup.manager import BackupManager
+
+__all__ = [
+    "ARCHIVE_QUALITIES",
+    "ARCHIVE_QUALITY_COMPRESSED",
+    "ARCHIVE_QUALITY_ORIGINAL",
+    "BACKUP_EXPORT_DIR",
+    "BACKUP_FORMAT_VERSION",
+    "BackupManager",
+    "RESTORE_LOCK_KEY",
+    "RESTORE_LOCK_TTL_SECONDS",
+    "RESTORE_STAGING_DIRNAME",
+]
