+++ backend/core/backup/restore/__init__.py
@@ -1,0 +1,1 @@
+"""Restoring an installation from a backup archive."""
